api/requirements copy.txt
async-generator==1.10
attrs==22.2.0
certifi @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_5d968ni_yn/croot/certifi_1671487774636/work/certifi
chromedriver-binary==110.0.5481.77.0
click @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_croot-quwchbn8/click_1646123324461/work
Deprecated==1.2.14
exceptiongroup==1.1.0
Flask==2.2.3
Flask-Limiter==3.5.0
gunicorn==20.1.0
h11==0.14.0
idna==3.4
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_70fph11sxu/croot/joblib_1666298846160/work
limits==3.6.0
markdown-it-py==3.0.0
MarkupSafe==2.1.2
mdurl==0.1.2
mysqlclient==2.1.1
numpy @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_d18vvv3mag/croot/numpy_and_numpy_base_1672336185310/work
ordered-set==4.1.0
outcome==1.2.0
packaging==23.2
Pygments==2.16.1
PySocks==1.7.1
python-dotenv @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_3ahu1_p9lb/croot/python-dotenv_1669132571851/work
rich==13.6.0
scikit-learn @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_05clcutezw/croot/scikit-learn_1673957225637/work
scipy==1.9.3
selenium==4.7.2
sniffio==1.3.0
sortedcontainers==2.4.0
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.8.0
urllib3==1.26.14
Werkzeug==2.2.3
wrapt==1.16.0
wsproto==1.2.0

